Added login page to project
I added the login page to the project, my special CSS file, and got it
to work. I'm also doing some work on test pages in the project to find
out more about how Bootstrap, etc. will work with Visual Studio. I'll
obviously need to remove the links as URLs to bootstrap and just use the
built-in one, but I'm leaving the URL for now.